+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [v0.0.17] - TBD
 
-Uses libvips v8.17.3, compiled with Emscripten v4.0.21.
+Uses libvips v8.18.0-rc2, compiled with Emscripten v4.0.21.
 
 ### Added
 
@@ -18,6 +18,7 @@ Uses libvips v8.17.3, compiled with Emscripten v4.0.21.
 - Switch from spng to libpng.
   [#105](https://github.com/kleisauke/wasm-vips/pull/105)
 - Make `filename` properties optional.
+- Update methods/enums for libvips 8.18.
 
 ## [v0.0.16] - 2025-11-07